+++ app/ui/components/sidebar/sidebar.styl
@@ -1,7 +1,7 @@
 .r-sidebar
-	width: 60px
+	width: 100px
 	height: 100%
-	padding-top: 20px
+	padding-top: 40px
 	overflow-y: scroll
 	overflow-x: hidden
 	text-align: center
@@ -9,3 +9,6 @@
 		width: 50px
 		height: 50px
 		border-radius: 50%
+	.menu-item
+		li
+			margin-top: 20px
