+++ app/ui/containers/test.js
@@ -1,19 +1,14 @@
 'use strict'
 
 import React, { Component } from 'react'
-import { Router, Route } from 'react-router'
-
-import Clock from 'com/clock/clock'
-import World from 'com/world/world'
-import Canvas from 'com/canvas/canvas'
 
 export default class Test extends Component {
-    constructor (props) {
+    constructor(props) {
         super(props)
     }
-    render () {
+    render() {
         return (
-            <h4>这是测试页面</h4>
+            <div>这是测试页面</div>
         )
     }
 }
