+++ app/ui/resource/css/common.styl
@@ -15,7 +15,7 @@
 	width: 100%
 	height: 100%
 .content-wrap
-	padding-left: 60px
+	padding-left: 100px
 	width: 100%
 	height: 100%
 	overflow: hidden
