+++ app/ui/resource/css/mixin.styl
@@ -14,3 +14,6 @@ ul
 	padding: 0
 	li
 		list-style-type: none
+
+a
+	text-decoration: none
